Changes by spackjarrow: "Update Docker container instructions and cleanup SQL script"
This commit updates the naming convention for Docker container instructions in readme.md file for consistency and ease of use. It also adds step-by-step instructions on how to use the shell for interacting with the MySQL and MongoDB containers, which will make it easier for other developers to access and use these resources. In the ITBA_2023_esquema_facturacion.sql file, redundant comments and white spaces were cleaned up for readability and efficiency.