Improve error handling in MySQL User API
This commit improves the error handling in the user API for MySQL. Now, instead of just logging the error and sending a 500 HTTP status, it checks if the error is related to the 'nro_cliente' duplicity by verifying error codes (1062, 1064, 1452). If such error happens, it returns a 400 status with a more descriptive message. Additionally, it has been ensured that a HTTP 200 status is only returned when 'sqlResult' is true, otherwise, a 404 status is sent.